refactor: move t helpers to utils
